dev: add ExampleSendEmailOrWebPushApi project (and abstractions)
To achieve an Example API the code needs to be in abstracted projects/libraries. There now exists two new projects: `Firepuma.EmailAndPush.Abstractions` and `Firepuma.EmailAndPush.Client`. The Abstractions contains shared models and basic logic, which is references/used by both the Client and FunctionApp. The Client project contains the code to send ServiceBus messages to the API/FunctionApp. Client code caters for validating the input DTO before sending it to ServiceBus. The service bus message is also decorated with `applicationId` in the `ApplicationProperties`.